+++ app/src/main/java/org/linphone/chat/ChatActivity.java
@@ -19,26 +19,34 @@
  */
 package org.linphone.chat;
 
+import android.app.Dialog;
 import android.app.Fragment;
 import android.app.FragmentManager;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Button;
+import android.widget.ImageView;
 import android.widget.Toast;
 import java.util.ArrayList;
+import org.linphone.LinphoneManager;
 import org.linphone.R;
 import org.linphone.activities.MainActivity;
 import org.linphone.contacts.ContactAddress;
 import org.linphone.core.Address;
+import org.linphone.core.ChatMessage;
+import org.linphone.core.ChatRoom;
 import org.linphone.core.Factory;
 import org.linphone.core.tools.Log;
 import org.linphone.utils.FileUtils;
+import org.linphone.utils.LinphoneUtils;
 
 public class ChatActivity extends MainActivity {
     public static final String NAME = "Chat";
 
     private String mSharedText, mSharedFiles;
+    private ChatMessage mForwardMessage;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -215,6 +223,8 @@ public class ChatActivity extends MainActivity {
         ChatMessagesFragment fragment = new ChatMessagesFragment();
         fragment.setArguments(extras);
         changeFragment(fragment, "Chat room", isChild);
+
+        showForwardDialog(localAddress, peerAddress);
     }
 
     public void showChatRoom(Address localAddress, Address peerAddress) {
@@ -301,5 +311,73 @@ public class ChatActivity extends MainActivity {
         GroupInfoFragment fragment = new GroupInfoFragment();
         fragment.setArguments(extras);
         changeFragment(fragment, "Chat room group info", true);
+    }
+
+    public void showChatRoomEphemeral(Address peerAddress) {
+        Bundle extras = new Bundle();
+        if (peerAddress != null) {
+            extras.putSerializable("RemoteSipUri", peerAddress.asStringUriOnly());
+        }
+        EphemeralFragment fragment = new EphemeralFragment();
+        fragment.setArguments(extras);
+        changeFragment(fragment, "Chat room ephemeral", true);
+    }
+
+    public void forwardMessage(ChatMessage message) {
+        Log.i("[Chat] Message forwarding enabled");
+        goBack();
+        mForwardMessage = message;
+        Toast.makeText(this, R.string.toast_choose_chat_room_for_sharing, Toast.LENGTH_LONG).show();
+    }
+
+    private void showForwardDialog(final Address localAddress, final Address peerAddress) {
+        if (mForwardMessage == null) return;
+
+        final Dialog dialog =
+                LinphoneUtils.getDialog(
+                        this, getString(R.string.chat_message_forward_confirmation_dialog));
+        dialog.findViewById(R.id.dialog_delete_button).setVisibility(View.GONE);
+
+        ImageView icon = dialog.findViewById(R.id.dialog_icon);
+        icon.setVisibility(View.VISIBLE);
+        icon.setImageResource(R.drawable.forward_message_dialog_default);
+
+        Button cancel = dialog.findViewById(R.id.dialog_cancel_button);
+        cancel.setOnClickListener(
+                new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
+                        mForwardMessage = null;
+                        dialog.dismiss();
+                    }
+                });
+
+        Button ok = dialog.findViewById(R.id.dialog_ok_button);
+        ok.setVisibility(View.VISIBLE);
+        ok.setText(getString(R.string.chat_message_context_menu_forward));
+        ok.setOnClickListener(
+                new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
+                        doMessageForwarding(localAddress, peerAddress);
+                        dialog.dismiss();
+                    }
+                });
+
+        dialog.show();
+    }
+
+    private void doMessageForwarding(Address localAddress, Address peerAddress) {
+        if (mForwardMessage != null) {
+            Log.i("[Chat] Found message to forward");
+            ChatRoom room = LinphoneManager.getCore().getChatRoom(peerAddress, localAddress);
+            if (room != null) {
+                Log.i("[Chat] Found chat room in which to forward message");
+                ChatMessage message = room.createForwardMessage(mForwardMessage);
+                message.send();
+                mForwardMessage = null;
+                Log.i("[Chat] Message forwarded");
+            }
+        }
     }
 }